% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/necklace.R
\name{divisors}
\alias{divisors}
\title{Find divisors}
\usage{
divisors(n)
}
\arguments{
\item{n}{Number to find divisors of}
}
\value{
The divisors
}
\description{
Find the integer divisors (including \code{1} and \code{n}) of \code{n}.
}
\author{
David Sterratt
}

