% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataset.R
\name{filter.table}
\alias{filter.table}
\title{Filter data frame}
\usage{
filter.table(tab, filt = list())
}
\arguments{
\item{tab}{Data frame to filter}

\item{filt}{List of values of table headings to filter on}
}
\value{
Filtered data frame
}
\description{
Filter data frame
}
\examples{
dat <- data.frame(a=1:5,
                  b=c(2, 2, 3, 3, 3),
                  c=c(5, 5, 5, 7, 7))
filter.table(dat, list(a=1))
filter.table(dat, list(b=3, c=5))
}
\author{
David Sterratt
}

