% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeps.R
\name{getDeps}
\alias{getDeps}
\title{Get all dependencies of package recursively}
\usage{
getDeps(pkg)
}
\arguments{
\item{pkg}{Name of package}
}
\value{
Vector of pakcage names and versions that \code{pkg} depends on
}
\author{
David Sterratt
}

