% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/necklace.R
\name{necklace}
\alias{necklace}
\title{Number of necklaces}
\usage{
necklace(k, n)
}
\arguments{
\item{k}{Number of symbols in the alphabet}

\item{n}{Length of string}
}
\value{
Number of necklaces
}
\description{
Compute the number of necklaces of length \code{n} formed from an
alphabet with \code{k} symbols. See
\url{https://en.wikipedia.org/wiki/Necklace_(combinatorics)}.
}
\author{
David Sterratt
}

