% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataset.R
\name{read.pars}
\alias{read.pars}
\title{Read parameter file}
\usage{
read.pars(id = NULL, file = "pars.dat")
}
\arguments{
\item{id}{Identity of parameter set in file to read}

\item{file}{Name of file}
}
\value{
If \code{id} is specified, return return parameters from
row of parameter table specified by \code{id}; otherwise return
whole table
}
\description{
Read a parameter file, in which each row contains a set of parameters.
}
\author{
David Sterratt
}

