% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataset.R
\name{write.dataset}
\alias{write.dataset}
\title{Write a dataset to file}
\usage{
write.dataset(x, dataset, suffix = ".dat")
}
\arguments{
\item{x}{Data frame containing dataset}

\item{dataset}{Name of dataset}

\item{suffix}{The suffix of the data file}
}
\description{
Write a dataset to file
}
\author{
David Sterratt
}

